pyproject.toml
[project]
name = "machine-learning-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "kagglehub>=0.3.13",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ratelimit>=2.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.45",
    "trafilatura>=2.0.0",
    "uvicorn>=0.40.0",
]
